requirements.txt
Django==1.8.6
https://static.dpaw.wa.gov.au/static/py/dpaw-utils/dist/dpaw-utils-0.3a3.tar.gz
Unipath==1.1
Fabric==1.10.2
jinja2==2.8
dj-static==0.0.6
django-debug-toolbar==1.4
django-crispy-forms==1.5.1
xlrd==0.9.3
xlutils==1.7.1
xlwt==0.7.5
webtemplate-dpaw==0.4.2
openpyxl==2.2.5
coverage==3.7.1
mixer==5.1.8
